requirements.txt
twisted
wheel
loguru
setuptools
black
pyopenssl
service_identity
pycryptodome
psycopg2-binary
psycopg2
peewee
sshtunnel
python-dotenv
peewee==3.14.8
peewee-migrations==0.3.30